@@ -14,8 +14,8 @@
     <name>SpringBot</name>
     <description>Telegram bot on Spring Boot</description>
     <properties>
-        <maven.compiler.source>17</maven.compiler.source>
-        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.source>19</maven.compiler.source>
+        <maven.compiler.target>19</maven.compiler.target>
         <telegram.version>5.6.0</telegram.version>
     </properties>
     <dependencies>
@@ -95,7 +95,7 @@
                 <configuration>
                     <executable>true</executable>
                 </configuration>
-            </plugin><plugin><groupId>org.apache.maven.plugins</groupId><artifactId>maven-compiler-plugin</artifactId><configuration><source>17</source><target>17</target></configuration></plugin>
+            </plugin><plugin><groupId>org.apache.maven.plugins</groupId><artifactId>maven-compiler-plugin</artifactId><configuration><source>19</source><target>19</target></configuration></plugin>
         </plugins>
     </build>
 
