@@ -10,7 +10,7 @@ import java.util.ResourceBundle;
  * @created 06.02.2023 14:53
  */
 public class LanguageLocalisation {
-    public static ResourceBundle messages = ResourceBundle.getBundle("messages", new Locale("ru"));
+    public static ResourceBundle messages = ResourceBundle.getBundle("messages", Locale.of("ru"));
     public static final String russianLang = EmojiParser.parseToUnicode("Русский:ru:");
     public static final String englishLang = EmojiParser.parseToUnicode("English:gb:");
 
